update wmb pkg tests
